Major code update 01.07.2024
Restarted work on finishing this project after more than a year.

Code was completely refactored. DMA/PIO interrupt was removed, and two state machines from the same PIO block were loaded with runup and (mock) residue code and sychronized using interrupt 0 (irq 0). Behaviour is exactly as expected. This is a huge step forward for this project.